--- src/styles/theme.scss
@@ -1,36 +1,0 @@
-@use 'sass:color';
-
-@use '@material/theme/color-palette';
-
-
-$primary: #006e54;
-$secondary: #a5d6a7;
-$surface: #fff;
-$background: #fff;
-$error: color-palette.$red-900;
-
-// Svelte Colors!
-@use '@material/theme/index' as theme with (
-  $primary: #006e54,
-  $secondary: #a5d6a7,
-  $surface: #fff,
-  $background: #fff,
-  $error: color-palette.$red-900,
-);
-
-html,
-body {
-  background-color: theme.$surface;
-  color: theme.$on-surface;
-}
-
-a {
-  color: #40b3ff;
-}
-a:visited {
-  color: color.scale(#40b3ff, $lightness: -35%);
-}
-
-* {
-  font-family: var(--font-family) !important;
-}
